recycler delete fix

+++ app/src/main/java/com/example/youtubebackgroundplayer/data/entity/VideoEntity.kt
@@ -3,24 +3,27 @@ package com.example.youtubebackgroundplayer.data.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.example.youtubebackgroundplayer.data.dto.VideoDto
-import com.example.youtubebackgroundplayer.util.Seconds
-import com.example.youtubebackgroundplayer.util.TimeUnit
+import com.example.youtubebackgroundplayer.util.time.Seconds
 
 @Entity(tableName="videos")
 data class VideoEntity (
     @PrimaryKey(autoGenerate = true) var id: Int,
     var title: String,
-    var duration: Long?,
+    var duration: Int?,
     var isWatched: Boolean,
     var videoId: String
 ) {
-    constructor(title: String, duration: Long?, isWatched: Boolean, videoId: String)
+    constructor(title: String, duration: Int?, isWatched: Boolean, videoId: String)
             : this(0, title, duration, isWatched, videoId)
 
     fun toVideoDto() =
         VideoDto(
             title,
-            duration?.let { Seconds(it) },
+            duration?.let {
+                Seconds(
+                    it
+                )
+            },
             isWatched,
             videoId
         )
